Moved cmdlets into runbits.
