--- pkg/cmdlets/README.md
@@ -1,3 +1,0 @@
-Commandlets (cmdlets) are packages that contain logic used by our commands that do not belong in any other package. In 
-most cases this logic is contained in the command itself, but in some cases logic needs to be shared between commands, 
-this is where cmdlets come in.
